+++ src/common/js/index.js
@@ -2,15 +2,15 @@
  * pwix:field/src/common/js/index.js
  */
 
-import { Def } from '../classes/def.class.js';
-import { Set } from '../classes/set.class.js';
-
 import './global.js';
 import './constants.js';
 import './trace.js';
 //
 import './configure.js';
 import './functions.js';
+
+import { Def } from '../classes/def.class.js';
+import { Set } from '../classes/set.class.js';
 
 Field.Def = Def;
 Field.Set = Set;